refactor(api): deprecate old HttpClient and introduce unified HTTP client for improved server and client handling
